include device_type when submitting the scorecard and move the scorecard attributes fuction to util file
